requirements.txt
asks
aioredis==1.3.0
Quart==0.15.1
Quart-Trio==0.8.0
trio==0.19.0
trio-asyncio==0.12.0
flake8
python-decouple